Move token calculation from OpenAIClient to OpenAITranslationAdapter
